Move metadata/labels helpers into system package
