+++ system/version_dump_windows.go
@@ -1,3 +1,5 @@
+// +build windows
+
 package system
 
 import (
@@ -7,6 +9,7 @@ import (
 	"github.com/buildkite/agent/logger"
 )
 
+// VersionDump returns a string representing the operating system
 func VersionDump(_ *logger.Logger) (string, error) {
 	dll := syscall.MustLoadDLL("kernel32.dll")
 	p := dll.MustFindProc("GetVersion")
